refactor(audio): simplify dispatch and extract wake word stripping
Replace the recording-path if/else with a match, inline the one-shot
chunk pipeline helper, move RecorderStartError to types, and split the
wake word stripping into clean_recording. Drop the self-addressed
long-dictation-segment event in favor of a direct call.
