+++ src-tauri/src/audio/audio.rs
@@ -1,18 +1,17 @@
 use crate::audio::helpers::{cleanup_recordings, ensure_recordings_dir, generate_unique_wav_name};
 use crate::audio::pipeline::process_recording;
 use crate::audio::recorder::AudioRecorder;
-use crate::audio::types::{AudioState, RecordingMode, RecordingTrigger};
+use crate::audio::clean_recording::strip_and_record;
+use crate::audio::types::{AudioState, RecorderStartError, RecordingMode, RecordingTrigger};
 use crate::audio::ChunkPipeline;
 use crate::clipboard;
 use crate::engine::ParakeetEngine;
 use crate::model::Model;
 use crate::overlay::overlay;
-use crate::wake_word::wake_word::normalize_text;
 use anyhow::Result;
 use log::{debug, error, info, warn};
 use std::sync::atomic::Ordering;
 use std::sync::Arc;
-use strsim::levenshtein;
 use tauri::{AppHandle, Emitter, Manager};
 
 pub fn record_audio(app: &AppHandle, mode: RecordingMode) {
@@ -21,48 +20,52 @@ pub fn record_audio(app: &AppHandle, mode: RecordingMode) {
     if state.get_recording_trigger() != RecordingTrigger::WakeWord {
         state.set_recording_trigger(RecordingTrigger::Keyboard);
     }
-    // Wake word listener stays active: validate/cancel words work during keyboard-triggered recording
 
     if matches!(mode, RecordingMode::Llm | RecordingMode::Command) {
         crate::llm::warmup_ollama_model_background(app);
     }
 
-    // Three exclusive recording paths:
-    // - Wake word: single-shot, no chunking, no long dictation.
-    // - Live Text Mode (Standard + toggle on): write each segment on silence.
-    // - Otherwise (keyboard): buffered chunking, preview drives the overlay.
-    let s = crate::settings::load_settings(app);
-    let live = s.long_dictation_enabled && mode == RecordingMode::Standard;
-    if state.get_recording_trigger() == RecordingTrigger::WakeWord {
-        state.long_dictation_active.store(false, Ordering::SeqCst);
-        *state.chunk_pipeline.lock() = None;
-    } else if live {
-        state.long_dictation_active.store(true, Ordering::SeqCst);
-        *state.chunk_pipeline.lock() = None;
-    } else {
-        state.long_dictation_active.store(false, Ordering::SeqCst);
-        start_chunk_pipeline(app, &state);
+    let settings = crate::settings::load_settings(app);
+    match state.get_recording_trigger() {
+        RecordingTrigger::WakeWord => {
+            state.long_dictation_active.store(false, Ordering::SeqCst);
+            *state.chunk_pipeline.lock() = None;
+        }
+        _ if settings.long_dictation_enabled && mode == RecordingMode::Standard => {
+            state.long_dictation_active.store(true, Ordering::SeqCst);
+            *state.chunk_pipeline.lock() = None;
+        }
+        _ => {
+            state.long_dictation_active.store(false, Ordering::SeqCst);
+            *state.chunk_pipeline.lock() = Some(ChunkPipeline::start(app));
+        }
     }
 
     internal_record_audio(app);
 }
 
-/// Buffers every chunk and writes the assembled text once at finalize. The
-/// real-time preview (or visualizer) drives the overlay, so no buffered overlay
-/// event is emitted here.
-fn start_chunk_pipeline(app: &AppHandle, state: &AudioState) {
-    let pipeline = ChunkPipeline::start(app);
-    *state.chunk_pipeline.lock() = Some(pipeline);
+fn internal_record_audio(app: &AppHandle) {
+    debug!("Starting audio recording...");
+    let state = app.state::<AudioState>();
+
+    crate::audio::sound::prewarm(app);
+
+    match start_new_recorder(app, true) {
+        Ok(sample_rate) => {
+            debug!("Recording started");
+            let s = crate::settings::load_settings(app);
+            if s.overlay_mode.as_str() == "recording" {
+                overlay::clear_pending_flash(app);
+                overlay::show_recording_overlay(app);
+            }
+            crate::overlay::tray::set_tray_recording(app);
+            crate::audio::streaming::start_streaming(app, &state, sample_rate);
+        }
+        Err(RecorderStartError::InitFailed) => notify_recording_error(app),
+        Err(_) => {}
+    }
 }
 
-enum RecorderStartError {
-    Busy,
-    DirUnavailable,
-    InitFailed,
-    StartFailed,
-}
-
-/// The caller must not already hold the recorder lock.
 fn start_new_recorder(app: &AppHandle, play_sound: bool) -> Result<u32, RecorderStartError> {
     let state = app.state::<AudioState>();
 
@@ -104,30 +107,6 @@ fn start_new_recorder(app: &AppHandle, play_sound: bool) -> Result<u32, Recorder
     *state.current_file_name.lock() = Some(file_name);
     *recorder_guard = Some(recorder);
     Ok(sample_rate)
-}
-
-fn internal_record_audio(app: &AppHandle) {
-    debug!("Starting audio recording...");
-    let state = app.state::<AudioState>();
-
-    crate::audio::sound::prewarm(app);
-
-    match start_new_recorder(app, true) {
-        Ok(sample_rate) => {
-            debug!("Recording started");
-            let s = crate::settings::load_settings(app);
-            if s.overlay_mode.as_str() == "recording" {
-                overlay::clear_pending_flash(app);
-                overlay::show_recording_overlay(app);
-            }
-            crate::overlay::tray::set_tray_recording(app);
-            crate::audio::streaming::start_streaming(app, &state, sample_rate);
-        }
-        // Only a failed device init pops the mic-error overlay, preserving the
-        // original behavior: a busy recorder or a failed stream start stays silent.
-        Err(RecorderStartError::InitFailed) => notify_recording_error(app),
-        Err(_) => {}
-    }
 }
 
 fn notify_recording_error(app: &AppHandle) {
@@ -189,17 +168,11 @@ pub fn stop_recording(app: &AppHandle) -> Option<std::path::PathBuf> {
             reset_recording_ui(app);
         }
     }
-
-    // Reset after processing so the last long-dictation segment is still
-    // formatted with the flag active (no short-text correction on it).
     state.long_dictation_active.store(false, Ordering::SeqCst);
 
     path
 }
 
-/// Chunked session: the worker buffered every chunk. Drain it for the stitched
-/// text, post-process once, then paste the whole block. The overlay is driven by
-/// the preview/visualizer, so no buffered/inserting/done event is emitted.
 fn finalize_chunked_session(
     app: &AppHandle,
     state: &AudioState,
@@ -240,23 +213,6 @@ fn finalize_single_recording(app: &AppHandle, state: &AudioState, path: &std::pa
             error!("Processing failed: {}", e);
             reset_recording_ui(app);
         }
-    }
-}
-
-/// Strips a trailing validation/submit wake word from the final concatenated
-/// text and keeps history in sync. Applied once, never per chunk.
-fn strip_and_record(app: &AppHandle, state: &AudioState, text: String) -> String {
-    match state.strip_word.lock().take() {
-        Some(word) => {
-            let stripped = strip_trailing_wake_word(&text, &word);
-            if stripped != text {
-                if let Err(e) = crate::history::update_last_transcription(app, stripped.clone()) {
-                    error!("Failed to update history after wake word strip: {}", e);
-                }
-            }
-            stripped
-        }
-        None => text,
     }
 }
 
@@ -308,7 +264,7 @@ pub fn cancel_recording(app: &AppHandle) {
     info!("Recording cancelled by user");
 }
 
-pub fn flush_and_continue_dictation(app: &AppHandle) {
+pub(super) fn flush_and_continue_dictation(app: &AppHandle) {
     let state = app.state::<AudioState>();
     if !state.long_dictation_active.load(Ordering::SeqCst) {
         return;
@@ -419,12 +375,8 @@ fn reset_recording_ui_delayed(app: &AppHandle, delay_ms: u64) {
 }
 
 pub fn write_transcription(app: &AppHandle, transcription: &str) -> Result<()> {
-    // Linux+Wayland only: hide BEFORE paste so KWin/Mutter hand
-    // keyboard focus back to the target app before Ctrl+V fires.
-    // The 400 ms settle in `clipboard::paste_with_delay` relies on
-    // this order via `overlay::millis_since_last_overlay_hide`.
-    // Other platforms keep the original timing (hide after paste in
-    // `reset_recording_ui*`).
+    // Wayland: hide before paste so KWin/Mutter returns focus before Ctrl+V.
+    // paste_with_delay's 400 ms settle relies on millis_since_last_overlay_hide.
     #[cfg(target_os = "linux")]
     {
         if crate::utils::platform::is_wayland_session() {
@@ -467,49 +419,6 @@ pub fn simulate_enter_key(app: &AppHandle) -> Result<(), String> {
     })
 }
 
-fn strip_trailing_wake_word(text: &str, wake_word: &str) -> String {
-    let ww = wake_word.trim();
-    if ww.is_empty() {
-        return text.to_string();
-    }
-
-    let trimmed = text.trim();
-    let text_words: Vec<&str> = trimmed.split_whitespace().collect();
-
-    let ww_normalized = normalize_text(ww);
-    let ww_words: Vec<&str> = ww_normalized.split_whitespace().collect();
-
-    if text_words.len() < ww_words.len() {
-        return trimmed.to_string();
-    }
-
-    // Search within the last words with a margin of 2 for trailing noise from STT
-    let margin = 2;
-    let earliest_start = text_words.len().saturating_sub(ww_words.len() + margin);
-
-    for start in earliest_start..=(text_words.len() - ww_words.len()) {
-        let candidate = &text_words[start..start + ww_words.len()];
-
-        let all_match = candidate.iter().zip(ww_words.iter()).all(|(tw, ww_w)| {
-            let tw_norm = normalize_text(tw);
-            let max_distance = if ww_w.len() <= 3 { 1 } else { 2 };
-            levenshtein(&tw_norm, ww_w) <= max_distance
-        });
-
-        if all_match {
-            // Remove everything from the matched position to the end
-            let result = text_words[..start].join(" ");
-            debug!(
-                "Stripped trailing wake word \"{}\" from transcription",
-                wake_word
-            );
-            return result;
-        }
-    }
-
-    trimmed.to_string()
-}
-
 pub fn write_last_transcription(app: &AppHandle, transcription: &str) -> Result<()> {
     if let Err(e) = clipboard::paste_last_transcript(transcription, app) {
         error!("Failed to paste last transcription: {}", e);
@@ -518,6 +427,7 @@ pub fn write_last_transcription(app: &AppHandle, transcription: &str) -> Result<
     debug!("Last transcription written to clipboard {}", transcription);
     Ok(())
 }
+
 
 pub fn preload_engine(app: &AppHandle) -> Result<()> {
     let state = app.state::<AudioState>();
@@ -539,107 +449,3 @@ pub fn preload_engine(app: &AppHandle) -> Result<()> {
     Ok(())
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn strip_exact_match_single_word() {
-        assert_eq!(
-            strip_trailing_wake_word("bonjour validate", "validate"),
-            "bonjour"
-        );
-    }
-
-    #[test]
-    fn strip_exact_match_multi_word() {
-        assert_eq!(
-            strip_trailing_wake_word("bonjour le monde alix validate", "alix validate"),
-            "bonjour le monde"
-        );
-    }
-
-    #[test]
-    fn strip_fuzzy_match_accent() {
-        // STT transcribes "validé" instead of "validate" — Levenshtein ≤ 2
-        assert_eq!(
-            strip_trailing_wake_word("bonjour alix validé", "alix validate"),
-            "bonjour"
-        );
-    }
-
-    #[test]
-    fn strip_fuzzy_match_typo() {
-        // STT transcribes "validatte" — Levenshtein ≤ 2
-        assert_eq!(
-            strip_trailing_wake_word("bonjour alix validatte", "alix validate"),
-            "bonjour"
-        );
-    }
-
-    #[test]
-    fn strip_fuzzy_match_missing_char() {
-        // STT transcribes "validat" — Levenshtein = 1
-        assert_eq!(
-            strip_trailing_wake_word("bonjour alix validat", "alix validate"),
-            "bonjour"
-        );
-    }
-
-    #[test]
-    fn strip_with_trailing_noise() {
-        // Trailing noise word after wake word — margin handles it
-        assert_eq!(
-            strip_trailing_wake_word("bonjour alix validate ok", "alix validate"),
-            "bonjour"
-        );
-    }
-
-    #[test]
-    fn strip_case_insensitive() {
-        assert_eq!(
-            strip_trailing_wake_word("bonjour Alix Validate", "alix validate"),
-            "bonjour"
-        );
-    }
-
-    #[test]
-    fn strip_no_match_returns_original() {
-        assert_eq!(
-            strip_trailing_wake_word("bonjour le monde", "alix validate"),
-            "bonjour le monde"
-        );
-    }
-
-    #[test]
-    fn strip_empty_wake_word() {
-        assert_eq!(
-            strip_trailing_wake_word("bonjour le monde", ""),
-            "bonjour le monde"
-        );
-    }
-
-    #[test]
-    fn strip_text_shorter_than_wake_word() {
-        assert_eq!(
-            strip_trailing_wake_word("validate", "alix validate"),
-            "validate"
-        );
-    }
-
-    #[test]
-    fn strip_only_wake_word() {
-        assert_eq!(
-            strip_trailing_wake_word("alix validate", "alix validate"),
-            ""
-        );
-    }
-
-    #[test]
-    fn strip_with_punctuation_from_stt() {
-        assert_eq!(
-            strip_trailing_wake_word("bonjour alix validate.", "alix validate"),
-            "bonjour"
-        );
-    }
-}
